Refactoring, added rotation for steering behaviors, some cleanup
