--- Assets/New Terrain 1.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 68894ee643ce0d94d9028efdaa928d61
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 15600000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
