--- Assets/New Terrain.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0e6b0295432a1f447856d37ca5261db0
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 15600000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
