--- Assets/Prefabs/FlockingPrefab.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: c6620a0f55343024789f088968c96f1e
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
