+++ Assets/Scenes/Nav/NavPerformanceDemo.unity
@@ -3428,7 +3428,7 @@ MonoBehaviour:
   priority: 0
   blendDistance: 0
   weight: 1
-  sharedProfile: {fileID: 11400000, guid: 150a15cfdbc36f686b9f6daf6c8bcfef, type: 2}
+  sharedProfile: {fileID: 11400000, guid: f45bc2788d2afa041ba167ccb57b2373, type: 2}
 --- !u!4 &524292823
 Transform:
   m_ObjectHideFlags: 0
