--- Packages/com.reese.nav/Agent/Systems/NavLerpSystem.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 80cc01a30aae4ba68978302b9c9da67f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
