+++ Packages/com.reese.nav/Agent/Systems/NavSteeringSystem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b9a44457e03b04e4bba7f13f585d52c1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
