+++ Packages/com.reese.nav/NavSystem.cs
@@ -29,7 +29,6 @@ namespace Reese.Nav
             SeparationWeight = 2.0f,
             AlignmentWeight = 1.5f,
             CohesionWeight = 0.1f,
-            WaypointWeight = 2.0f,
             FollowWeight = 2.5f,
             CollisionAvoidanceStrength = 0.5f,
             QuadrantCellSize = 5,
@@ -85,11 +84,25 @@ namespace Reese.Nav
         /// <summary>Includes settings used by the navigation/ flocking systems.</summary>
         public struct NavFlockingSettings
         {
+            /// <summary>
+            /// Works like collision, so that units do not share the same space
+            /// </summary>
             public float SeparationWeight;
+            /// <summary>
+            /// How much should agents align with other agents in a flock
+            /// </summary>
             public float AlignmentWeight;
+            /// <summary>
+            /// How strongly should agents look to join other agents (grouping behavior)
+            /// </summary>
             public float CohesionWeight;
-            public float WaypointWeight;
-            public float FollowWeight;
+            /// <summary>
+            /// To be included, maybe rewrite the NavFollow system to add weighted steering to the NavSteeringSystem?
+            /// </summary>
+            public float FollowWeight; 
+            /// <summary>
+            /// Weight of collision avoidance so agents steer away from each other
+            /// </summary>
             public float CollisionAvoidanceStrength;
             public int QuadrantCellSize; 
             public int QuadrantZMultiplier;
