+++ UserSettings/EditorUserSettings.asset
@@ -24,19 +24,19 @@ EditorUserSettings:
       value: 22424703114646680e0b0227036c6c040416162e292c68252320092a
       flags: 0
     RecentlyUsedScenePath-4:
-      value: 22424703114646680e0b0227036c71110058362b3a05292624271a19f7ed241ce7ea39a7f234362820
-      flags: 0
-    RecentlyUsedScenePath-5:
       value: 22424703114646680e0b0227036c71110058362b3a003f323f201917e7ed3b76f7e93ffdfe
       flags: 0
-    RecentlyUsedScenePath-6:
+    RecentlyUsedScenePath-5:
       value: 22424703114646680e0b0227036c71110058362b3a182939233d3c3de6c33831e1ec12ecea357129370bfb25
       flags: 0
+    RecentlyUsedScenePath-6:
+      value: 22424703114646680e0b0227036c71110058362b3a05292624271a19f7ed241ce7ea39a7f234362820
+      flags: 0
     RecentlyUsedScenePath-7:
-      value: 22424703114646680e0b0227036c71110058362b3a1823222b260f3ee3ee373dc6e23be6a92f31352d1b
+      value: 22424703114646680e0b0227036c71110058362b3a1c23223f28143dc6e53937acf238e0f323
       flags: 0
     RecentlyUsedScenePath-8:
-      value: 22424703114646680e0b0227036c71110058362b3a1c23223f28143dc6e53937acf238e0f323
+      value: 22424703114646680e0b0227036c71110058362b3a1823222b260f3ee3ee373dc6e23be6a92f31352d1b
       flags: 0
     RecentlyUsedScenePath-9:
       value: 22424703114646680e0b0227036c71110058362b3a0e2a3f2e22143de5c43135eda923e7ee2e26
